change(Calculator) add method's return type from int to long

+++ src/main/java/me/rubaiya/Calculator.java
@@ -1,7 +1,7 @@
 package me.rubaiya;
 
 public class Calculator {
-    public int add (int a, int b) {
+    public long add (int a, int b) {
         return a + b;
     }
 }
